% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adding_data_fns.R
\name{add_dataframe}
\alias{add_dataframe}
\title{Materialize a dataframe to Stardog}
\usage{
add_dataframe(stardog, df, mapping, graph = NULL, verbose = FALSE)
}
\arguments{
\item{stardog}{Stardog object}

\item{df}{Dataframe with the data to be imported to Stardog}

\item{mapping}{Mapping function expressed as a string}

\item{graph}{optional named of named graph to receive the data}

\item{verbose}{if TRUE, get more details about the request}
}
\description{
Materialize a dataframe to Stardog
}
