% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_db_metadata_fns.R
\name{add_namespace}
\alias{add_namespace}
\title{Add a namespace to the database}
\usage{
add_namespace(stardog, uri = "http://iris.com/", prefix = "iris")
}
\arguments{
\item{stardog}{stardog connection object}

\item{uri}{Full uri of the namespace}

\item{prefix}{The prefix for that uri}
}
\value{
Does nothing if the namespace already exists. If not, adds the namespace/prefix pair to the existing namespace
}
\description{
Add a namespace to the database
}
\details{
Checks to see if the namespace already exists.
}
