% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stored_query_fns.R
\name{add_stored}
\alias{add_stored}
\title{Store a named query}
\usage{
add_stored(
  stardog,
  q = "select (count(*) as ?n) {?s ?p ?o .}",
  queryName = "CountEm",
  reasoning = FALSE
)
}
\arguments{
\item{stardog}{Stardog object}

\item{q}{Query to be stored}

\item{queryName}{Name for the stored query}

\item{reasoning}{Boolean. If TRUE, reasoning is used in the query.}
}
\value{
Returns a success message or the status code of the call
}
\description{
Store a named query
}
