% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adding_data_fns.R
\name{add_ttl}
\alias{add_ttl}
\title{Add a string of turtle}
\usage{
add_ttl(stardog, ttl = NULL, graph = NULL, path = TRUE)
}
\arguments{
\item{stardog}{Stardog object}

\item{ttl}{Some lines of turtle in a string.}

\item{graph}{optional. name of the named graph to receive the new data}

\item{path}{boolean. If true, then ttl refers to the path to the ttl file.}
}
\value{
nothing. A message is issued to indicate success or failure. Upon failure
it returns the status code
}
\description{
Supply RDF in turtle format as a (possibly) multi-line string
}
