% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virtual_graphs.R
\name{add_virtual_graph}
\alias{add_virtual_graph}
\title{Create a virtual graph}
\usage{
add_virtual_graph(stardog, vg_name, source_name, mappings = "", update = FALSE)
}
\arguments{
\item{stardog}{Stardog object}

\item{vg_name}{Name of the virtual graph}

\item{source_name}{name of the data source}

\item{mappings}{SMS mappings as a character string}

\item{update}{boolean, if TRUE then update an existing VG graph}
}
\value{
success code of the Post
}
\description{
Set up a virtual graph against a registered data source. Supply
the mappings as a character string.
}
