% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_fns.R
\name{ask}
\alias{ask}
\title{Issues an ask query to Stardog and return the result as a boolean}
\usage{
ask(stardog, q = NA, graph = NA, ...)
}
\arguments{
\item{stardog}{A stardog object for the connect, containing a database}

\item{q}{The query expressed as a string. Must be an ask query}

\item{graph}{Run the query against this named graph}

\item{...}{Options for the query. See Details}
}
\value{
boolean
}
\description{
Issues an ask query to Stardog and return the result as a boolean
}
\details{
Additional parameters:

- limit (integer) number of results to return

- reasoning (boolean) turn on reasoning. Not needed if a schema is included.

- schema (string) name of the reasoning schema

- offset (integer) number of results to skip before returning rows.
}
