% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beaches.R
\docType{data}
\name{beaches}
\alias{beaches}
\title{The Toronto Beaches data}
\format{
An object of class \code{data.frame} with 7269 rows and 7 columns.
}
\source{
City of Toronto public data portal
}
\usage{
beaches
}
\description{
This set contains daily eColi readings over several summers
taken as Toronto's public beaches.
}
\keyword{datasets}
