% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{begin_tx}
\alias{begin_tx}
\title{Start a transaction}
\usage{
begin_tx(stardog)
}
\arguments{
\item{stardog}{Stardog object}
}
\value{
Stardog object with the tx id in the transaction field.
}
\description{
Start a transaction
}
\details{
Some operations on the Stardog database require that a transaction be initiated. A transaction
id is generated, which can be used to roll back the transaction should a problem arise.
}
