% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reasoning_schema_fns.R
\name{build_model_list}
\alias{build_model_list}
\title{Internal function to prepare a list of models}
\usage{
build_model_list(model_list, old_models = NULL)
}
\arguments{
\item{model_list}{Named list}

\item{old_models}{Named list}
}
\value{
list with one element called `reasoning.names`. Its contents describe one or more models. Each model is named by
the name it will have in the stardog database. The contents of the model are a vector of graph uri's. These graphs defined
the model.
}
\description{
Prepare a list of model names with associated graphs to be sent as
part of the body in a request for setting or adding to the list of models
}
