% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starmap.R
\name{buildText}
\alias{buildText}
\title{build the sparql and onto text for one formula}
\usage{
buildText(formula, data, holding)
}
\arguments{
\item{formula}{single formula (string)}

\item{data}{The dataframe to be parsed}

\item{holding}{A list of text fragments thus far}
}
\value{
An updated version of the holding list
}
\description{
build the sparql and onto text for one formula
}
