% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starmap.R
\name{buildTextSide}
\alias{buildTextSide}
\title{build the sparql and onto text for a side of a formula}
\usage{
buildTextSide(side, data, holding)
}
\arguments{
\item{side}{the formula fragment (side) being parsed here}

\item{data}{the dataframe to be parsed}

\item{holding}{A list containing all the interim calculations}
}
\value{
a list of text to be added to the sms and onto files and the nodeList
}
\description{
build the sparql and onto text for a side of a formula
}
\details{
updates the nodeList with new nodes. This allows the same field to appear in different
formulae without getting extra BIND statements in the where clause, or extra definitions
in the ontology file. Updates the text for the sparql clause, the bindings and the ontology
All of this is held in the holding list.

The side parameter is a list of two elements: lhs and rhs. lhs is always a single token. rhs could be a single token
or a vector of tokens.

This function does not build on mapping fragments already constructed. It builds its own contribution, which are combined
by the calling functions to build the complete mapping and ontology files for all of formulae being parsed.
}
