% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starmap.R
\name{capitalize}
\alias{capitalize}
\title{Capitalize the first letter of a string, make the rest lower case}
\usage{
capitalize(var)
}
\arguments{
\item{var}{a string}
}
\value{
The string in title case
}
\description{
Used to convert variable names to camel case or pascal case.
}
