% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{check_db}
\alias{check_db}
\title{check if the database field is empty}
\usage{
check_db(stardog, silent = TRUE)
}
\arguments{
\item{stardog}{Stardog object}

\item{silent}{if TRUE, suppress output.}
}
\value{
Returns a Boolean.
}
\description{
If the database field is empty, returns a message (on silent = FALSE) to suggest running \code{use_database} to set the database name.
Returns TRUE or FALSE according as the name field holds a name or not. Does not check for the actual existence of the database.
}
