% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starmap.R
\name{checkField}
\alias{checkField}
\title{Test if a string is a field name of the database}
\usage{
checkField(field, dfNames)
}
\arguments{
\item{field}{The test string}

\item{dfNames}{The field names of the database being mapped}
}
\value{
Boolean. TRUE if the field is a name of the database
}
\description{
Simple function to save some typing
}
\details{
Legitimate fields are either names of the dataframe, or \code{#var}, where
we want a rownumber to be added to the mapping with name var.
}
