% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{commit_tx}
\alias{commit_tx}
\title{Commit a transaction}
\usage{
commit_tx(stardog)
}
\arguments{
\item{stardog}{Stardog object}
}
\value{
stardog object, with transaction set to NA.
}
\description{
Commit a transaction
}
\details{
This function is part of the transaction sequence. It issues a request to commit
the transaction whose id is in the Stardog object. If successful, the transaction ID
is set to NA
}
