% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stored_query_fns.R
\name{delete_stored}
\alias{delete_stored}
\title{Delete a stored query}
\usage{
delete_stored(stardog, storedQuery = NULL, all = FALSE)
}
\arguments{
\item{stardog}{Stardog object}

\item{storedQuery}{Name of the stored query to be removed}

\item{all}{If True, delete all stored queries.}
}
\description{
Delete a stored query
}
\details{
Either set all=TRUE, or supply the name of a query.
}
