% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virtual_graphs.R
\name{delete_virtual_graph}
\alias{delete_virtual_graph}
\title{delete a virtual graph}
\usage{
delete_virtual_graph(stardog, vgName)
}
\arguments{
\item{stardog}{stardog object}

\item{vgName}{name of the virtual graph to be deleted}
}
\value{
Status code of the deletion
}
\description{
delete a virtual graph
}
