% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{df_to_string}
\alias{df_to_string}
\title{Converts a data frame to a string}
\usage{
df_to_string(x)
}
\arguments{
\item{x}{The data frame to be converted}
}
\value{
A string serialization of the dataframe
}
\description{
takes a data frame and converts to a serialized CSV string. To be used
by the add_dataframe function. No error checking takes place.
}
