% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_management_fns.R
\name{drop_database}
\alias{drop_database}
\title{Drops the named database}
\usage{
drop_database(stardog, silent = FALSE)
}
\arguments{
\item{stardog}{A stardog object}

\item{silent}{(boolean) If TRUE, suppress output}
}
\value{
Stardog object with NA in database slot.
}
\description{
Drops the database contained in the stardog object. If silent = FALSE, gives messages of
failure or success.
}
