% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virtual_graphs.R
\name{exists_data_source}
\alias{exists_data_source}
\title{Does a data source exist}
\usage{
exists_data_source(stardog, data_source)
}
\arguments{
\item{stardog}{Stardog object}

\item{data_source}{Name of the potential data source}
}
\value{
Boolean. True if there is a data source with that name
}
\description{
Checks if there is a data source with the supplied name
}
