% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_management_fns.R
\name{exists_database}
\alias{exists_database}
\title{check if there exists a database with the given name}
\usage{
exists_database(stardog, dbName)
}
\arguments{
\item{stardog}{Stardog object}

\item{dbName}{potential database name}
}
\value{
Boolean. True if there is a database with the supplied name
}
\description{
check if there exists a database with the given name
}
