% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virtual_graphs.R
\name{exists_virtual_graph}
\alias{exists_virtual_graph}
\title{does a virtual graph exist}
\usage{
exists_virtual_graph(stardog, vgName)
}
\arguments{
\item{stardog}{stardog object}

\item{vgName}{name of possibly existing virtual graph}
}
\value{
Boolean. True if there is a virtual graph with that namedd
}
\description{
Checks if a particular graph name corresponds to an existing virtual
graph.
}
