% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_db_metadata_fns.R
\name{fix_namespace}
\alias{fix_namespace}
\title{Internal function to replace urn's with prefix's}
\usage{
fix_namespace(x)
}
\arguments{
\item{x}{Content to be fixed}
}
\value{
A data frame with the prefix and URN's of the namespaces
}
\description{
Internal function to replace urn's with prefix's
}
\details{
An internal utility function to massage the output from the API and
return a nice dataframe of prefixes and URN's.
}
