% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{fix_results}
\alias{fix_results}
\title{Extract results of a GET query from the returned value}
\usage{
fix_results(r)
}
\arguments{
\item{r}{the response value from a Get.}
}
\value{
Dataframe containing the results of the query.
}
\description{
This function is used internally by the query function
}
\details{
R utility `type.convert` is used to convert the columns to the most likely datatypes
Typically, it will do a good job on integers, decimals and Booleans. Dates are
returned as character strings
}
