% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{fix_results_ask}
\alias{fix_results_ask}
\title{Extract results of an ASK query from the returned content}
\usage{
fix_results_ask(r)
}
\arguments{
\item{r}{the response value from a Get.}
}
\value{
Boolean containing the results of the ASK query
}
\description{
This function is used internally by the query function
}
