% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{format_delim_single_thread}
\alias{format_delim_single_thread}
\title{Turn a dataframe into a string}
\usage{
format_delim_single_thread(
  x,
  delim,
  na = NA,
  quote = c("needed", "all", "none"),
  escape = c("double", "backslash", "none"),
  eol = "\\n"
)
}
\arguments{
\item{x}{The data frame}

\item{delim}{the delimiter of the resulting delimited string}

\item{na}{How missing values are encoded}

\item{quote}{presumably the defaults are OK}

\item{escape}{likewise}

\item{eol}{end of line marker to include in the string to mark distinct rows of the data frame.}
}
\description{
This is a slight variation on the readr format_delim function. It forces the function
To single thread.
}
