% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reasoning_schema_fns.R
\name{generate_model}
\alias{generate_model}
\title{Generate a model from the graph data}
\usage{
generate_model(
  stardog,
  reasoning = FALSE,
  output = c("text", "owl", "sql", "graphql", "shacl"),
  encoding = "UTF-8"
)
}
\arguments{
\item{stardog}{Stardog object}

\item{reasoning}{Boolean}

\item{output}{output format. Choose from the list.}

\item{encoding}{Supply encoding for the response. Defaults to UTF-8}
}
\value{
Status code
}
\description{
Not really sure what this does and how it is relevant
}
