% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reasoning_schema_fns.R
\name{get_model_list}
\alias{get_model_list}
\title{Get the models used by the database}
\usage{
get_model_list(stardog)
}
\arguments{
\item{stardog}{Stardog object}
}
\value{
Vector of model names associated with this database.
}
\description{
Get the models used by the database
}
