% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_db_metadata_fns.R
\name{get_option}
\alias{get_option}
\title{Get a named option}
\usage{
get_option(stardog, option = "search.version")
}
\arguments{
\item{stardog}{Stardog connection object}

\item{option}{The name of the option. Defaults to search.version}
}
\value{
Returns the value of the named option.
}
\description{
Get a named option
}
