% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_db_metadata_fns.R
\name{get_options}
\alias{get_options}
\title{List database options}
\usage{
get_options(stardog)
}
\arguments{
\item{stardog}{stardog connection object}
}
\value{
Returns the options associated with the stardog database.
}
\description{
List database options
}
