% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stored_query_fns.R
\name{get_stored}
\alias{get_stored}
\title{Returns the stored query}
\usage{
get_stored(stardog, queryName = "CountEm", encoding = "UTF-8")
}
\arguments{
\item{stardog}{Stardog object}

\item{queryName}{Name of the stored query}

\item{encoding}{Defaults to UTF-8}
}
\value{
The content of the stored query.
}
\description{
Returns a string containing the script of the stored query
}
\details{
This allows you to inspect a stored query and read the sparql.
}
