% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starmap.R
\name{getTokenRange}
\alias{getTokenRange}
\title{Returi the token}
\usage{
getTokenRange(token, dfNames)
}
\arguments{
\item{token}{The token (String)}

\item{dfNames}{The names of the data frame being mapped}
}
\value{
A vector of field names corresponding to the token
}
\description{
If the token is a single field, that field is returned. If the token is
a range of integers or strings, it returns a list of field names within
that range
It checks if the range is legal and if the field names exist in the
database.
}
\details{
\code{1:4} returns fields 1 through 4 inclusive.
\code{var2} returns field var2
\code{var2:var4} returns fields var2, var3, var4.
\code{0:4} gives a warning and indexes from 1
If the upper bound of the range exceeds the number of fields, the function
issues a warning and takes everything from the lower limit on up. This is a compromise with
current R practice for slicing, which does not throw an out of bounds error, but slices to the
end of the vector and then fills in the extra space with NA.
}
