% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{goodBoy}
\alias{goodBoy}
\title{A fluff function to check the connection.}
\usage{
goodBoy(stardog)
}
\arguments{
\item{stardog}{a Stardog object}
}
\value{
"Woof!" if the connection is active; otherwise returns the HTTP response.
}
\description{
A fluff function to check the connection.
}
