% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{iri_to_prefix}
\alias{iri_to_prefix}
\title{changes iri nodes with complete IRI bases to prefix: format}
\usage{
iri_to_prefix(x, stardog)
}
\arguments{
\item{x}{a vector of results from a query, potentially containing iri bases}

\item{stardog}{the stardog object.}
}
\value{
Returns a transformed vector if all of the values have partial matches
in the namespace file. If some cannot be matched because, say, they are literals, then
it returns the vector unchanged
}
\description{
This function is intended to adjust query results to use the namespace
prefixes instead of complete IRI's. Typically, this function will be applied
to a vector of results. We can't assume that every item in the vector will have
the same base IRI and prefix, given the optional keyboard, but typically,
the prefix will be the same.
}
\details{
Note that this is the only function here in which the stardog object is in second place. That
is because the primary application of this function is as applied to the dataframe output
of a select query, and column vector has to be in first position.
}
