% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iris_df.R
\docType{data}
\name{iris_df}
\alias{iris_df}
\title{The Fisher Iris data}
\format{
An object of class \code{data.frame} with 150 rows and 7 columns.
}
\source{
Base R datasets.
}
\usage{
iris_df
}
\description{
This is the Fisher iris data that comes with R, only with the header
names changed (to get rid of '.') and with a row id added.
}
\keyword{datasets}
