% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virtual_graphs.R
\name{list_data_sources}
\alias{list_data_sources}
\title{List data sources}
\usage{
list_data_sources(stardog)
}
\arguments{
\item{stardog}{Stardog object}
}
\value{
a list of data sources
}
\description{
List data sources
}
