% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_management_fns.R
\name{list_databases}
\alias{list_databases}
\title{List the databases available on the Stardog instance}
\usage{
list_databases(stardog)
}
\arguments{
\item{stardog}{Stardog object}
}
\value{
Returns a vector of database names
}
\description{
List the databases available on the Stardog instance
}
