% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_graphs.R
\name{list_graphs}
\alias{list_graphs}
\title{List all named graphs in the database}
\usage{
list_graphs(stardog)
}
\arguments{
\item{stardog}{Stardog object}
}
\value{
Vector of named graphs
}
\description{
List all named graphs in the database
}
