% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stored_query_fns.R
\name{list_stored}
\alias{list_stored}
\title{List all stored queries}
\usage{
list_stored(stardog, allDB = FALSE)
}
\arguments{
\item{stardog}{Stardog object}

\item{allDB}{Boolean. Under FALSE (default), returns the queries for the currenct database}
}
\description{
Lists all stored queries, or only those available to the current database
}
\details{
If allDB is TRUE, then the function returns a database with the query names and their respective databases

 If alDB is FALSE, it returns a vector of the query names for the current database

 If the current database is NA, it returns all stored queries and sends a message.
}
