% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virtual_graphs.R
\name{list_virtual_graphs}
\alias{list_virtual_graphs}
\title{list virtual graphs}
\usage{
list_virtual_graphs(stardog)
}
\arguments{
\item{stardog}{stardog object}
}
\value{
the results
}
\description{
List the virtual graphs from the instance. It does not appear to allow you
to specify the database.
}
