% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starmap.R
\name{lowerCaseFirst}
\alias{lowerCaseFirst}
\title{Make the first letter of the string lower case}
\usage{
lowerCaseFirst(var)
}
\arguments{
\item{var}{a string}
}
\value{
the string with the first character in lower case, leaving the rest as is
}
\description{
Make the first letter of the string lower case
}
