% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_management_fns.R
\name{new_database}
\alias{new_database}
\title{Creates a new database}
\usage{
new_database(stardog, dbName)
}
\arguments{
\item{stardog}{Stardog object}

\item{dbName}{Name of the new database (string)}
}
\value{
The stardog object with dbName added to the database slot.
}
\description{
Checks to see if the named database already exists. If it exists already.
If it does not, then creates a new database. See replace_database to replace
an existing database.
}
