% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{parse_options}
\alias{parse_options}
\title{Parse optional parameters}
\usage{
parse_options(params, optionFun)
}
\arguments{
\item{params}{Named list of optional parameters}

\item{optionFun}{Function to supply the list of allowed parameters}
}
\value{
Boolean. TRUE if the parameter names are all in the allowed list.
}
\description{
Parse optional parameters
}
