% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotSchema}
\alias{plotSchema}
\title{plot the ontology based on a named graph}
\usage{
plotSchema(
  stardog,
  onto_graph,
  datatypes = FALSE,
  plot = TRUE,
  datatypeWt = 3,
  classColour = "skyblue",
  dataColour = "lightgreen"
)
}
\arguments{
\item{stardog}{Stardog object}

\item{onto_graph}{URI of the named graph}

\item{datatypes}{Boolean: if TRUE, include datatype properties.}

\item{plot}{Boolean: plot the graph on TRUE}

\item{datatypeWt}{Large values will pull the datatype properties closer to their parent node.}

\item{classColour}{Colour for the class nodes}

\item{dataColour}{Colour of the datatype nodes}
}
\value{
the igraph object if PLOT=FALSE. Otherwise, plots the schema
}
\description{
Plots the ontology, based solely on classes, or classes and
datatypes. Requires the ontology to be in a named graph.
TODO: allow used of the default graph.
}
