% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{prefix_to_iri}
\alias{prefix_to_iri}
\title{Convert a IRI with prefix to the full IRI}
\usage{
prefix_to_iri(stardog, s)
}
\arguments{
\item{stardog}{The stardog object}

\item{s}{the string to be converted}
}
\value{
the converted string, or simply s itself if no prefix was found in the namespace
}
\description{
If the string contains a single :, the functions checks if the first component
is a namespace prefix. If it is, then it concerts the string to complete format.
For example, if ex is the prefix for http://example.com/, it will convert
ex:myGraph to http://example.com/myGraph.
}
\details{
This is intended to parse named graphs entered as parameters. The function is not vectorized
and adjusts only one string in each call.
}
