% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{print.stardog}
\alias{print.stardog}
\title{print method for the Stardog object}
\usage{
\method{print}{stardog}(x, ...)
}
\arguments{
\item{x}{Stardog object to be printed out}

\item{...}{not actually used}
}
\description{
Only prints out the endpoint and the database, if there is one
Does not reveal username, password or variables used by stardogR
to keep track of things.
}
