% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starmap.R
\name{processFormula}
\alias{processFormula}
\title{process a formula}
\usage{
processFormula(f, dfNames)
}
\arguments{
\item{f}{a formula string}

\item{dfNames}{Names of the fields in the dataframe to be mapped}
}
\value{
formula parsed as a list of lists.
}
\description{
Processes an individual formula and convert to a list of lists
}
