% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starmap.R
\name{processPartial}
\alias{processPartial}
\title{Process the left or right side of the formula.}
\usage{
processPartial(snippet, dfNames)
}
\arguments{
\item{snippet}{string containing the right or left side of the formula}

\item{dfNames}{The data frame being mapped}
}
\value{
A list containing the left and right sides of the snippet.
}
\description{
Process the left or right side of the formula.
}
\details{
We're looking at snippets such as "x | y" or "x | y + z + 3:5"
}
