% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starmap.R
\name{processPlus}
\alias{processPlus}
\title{Process a formula with + in it}
\usage{
processPlus(snippet, dfNames)
}
\arguments{
\item{snippet}{The formula snippet}

\item{dfNames}{The names of the data frame being mapped}
}
\value{
A list of field names from the dataframe
}
\description{
Process formula sippets like \code{var1 + var2 + 4:6}
Calls getTokenRange for each term.
}
