% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_fns.R
\name{query}
\alias{query}
\title{Wrapper for specific query functions ask, construct and select}
\usage{
query(sg, q = "select (count(*) as ?n) {?s ?p ?o .}", namedGraph = NA, ...)
}
\arguments{
\item{sg}{Stardog object}

\item{q}{An ask, construct or select query}

\item{namedGraph}{runs the query against the specified named graph}

\item{...}{Options for the query. See functions ask, construct and select}
}
\value{
the results of the query
}
\description{
Detects the type of query and runs the appropriate function
}
