% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{query_options}
\alias{query_options}
\title{Query options}
\usage{
query_options()
}
\value{
Vector of parameters: reasoning, schema, timeout, graph-uri, offset and limit.
}
\description{
Options for informational queries
}
\details{
This is a wheeze for providing some constants to the rest of stardogR. These are the options
allowed by certain functions. The output of query_options can be used to check if the
supplied parameters fall into the allowed list.

It is available for export so that users can see the list of currently implemented
parameter options. There should be no other reason for invoking it.
}
