% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virtual_graphs.R
\name{refresh_metadata}
\alias{refresh_metadata}
\title{refresh the metadata}
\usage{
refresh_metadata(stardog, sourceName)
}
\arguments{
\item{stardog}{stardog object}

\item{sourceName}{name of the data source to be refreshed}
}
\value{
the status code of the post operation
}
\description{
If the tables of interest in a data source change, we need
to refresh the metadata, or Stardog will not be able to find
the new tables.
}
