% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namespaces.R
\name{remove_namespace}
\alias{remove_namespace}
\title{Removes a namespace}
\usage{
remove_namespace(stardog, prefix)
}
\arguments{
\item{stardog}{Stardog connection object}

\item{prefix}{The prefix of the namespace}
}
\value{
Returns the return value of the post operation
}
\description{
Removes a namespace/prefix pair upon supplying a prefix
}
