% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stored.R
\name{rename_stored}
\alias{rename_stored}
\title{Rename a stored query}
\usage{
rename_stored(stardog, oldName = "CountEm", newName = "foo")
}
\arguments{
\item{stardog}{Stardog object}

\item{oldName}{current name of stored query}

\item{newName}{rename the query to newName}
}
\value{
Success message
}
\description{
Rename a stored query
}
