% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_management_fns.R
\name{reset_database}
\alias{reset_database}
\title{Replace a database}
\usage{
reset_database(stardog)
}
\arguments{
\item{stardog}{A stardog object, including the name of the database to be reset}
}
\value{
Returns the stardog object, having recreated the database
}
\description{
Drops a database and creates a fresh one with the same name. All data
will be deleted.
}
