% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{rollback_tx}
\alias{rollback_tx}
\title{Roll back a transaction}
\usage{
rollback_tx(stardog)
}
\arguments{
\item{stardog}{Stardog object}
}
\value{
Stardog object with the transaction id set to NA. Rolls back the transaction
}
\description{
Roll back a transaction
}
