% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{sampleGraph}
\alias{sampleGraph}
\title{obtain a "random" sample from the graph}
\usage{
sampleGraph(stardog, triple, size = 100, graph = NA, ...)
}
\arguments{
\item{stardog}{a stardog object}

\item{triple}{a RDF triple pattern that forms the basis of the sample}

\item{size}{The number of triples in the random sample}

\item{graph}{the named graph where the sample is saved.}

\item{...}{The usual query parameters. For example, reasoning can be used.}
}
\value{
a data frame of the sampled triples.
}
\description{
Supply a simple pattern and pull out a random sample.
The entities thus selected may be saved to a named graph
With logging, the sample is connected to a sample node.
}
\details{
The sample is only saved if graph is present. Logging adds some additional structure
to the graph where the sample is saved.

This function invokes the Stardog sampling service. Read the Stardog documentation to learn more about the
probabilistic properties of this service. Note that sampling cannot be done on complex queries, but only
against simple triples. Pattern \code{?s a owl:someThing} would draw a sample of nodes of class someThing.
}
