% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reasoning_schema_fns.R
\name{set_models}
\alias{set_models}
\alias{clear_models}
\alias{remove_model}
\alias{add_models}
\title{Add models to the database}
\usage{
set_models(stardog, model_list)

clear_models(stardog)

remove_model(stardog, model_list)

add_models(stardog, model_list)
}
\arguments{
\item{stardog}{Stardog object}

\item{model_list}{Named list containing the model information.}
}
\value{
Success message from the response.
}
\description{
This function deletes all previous models and replaces them by the models
in the new list.
}
\details{
Each model is defined by one or more graphs that contain reasoning ontologies.
The names of the model list are passed to Stardog as the names of the reasoning schemas.
The graphs URI's must be given in full format without prefixes.
}
\section{Functions}{
\itemize{
\item \code{clear_models()}: Clears all models from the database

\item \code{remove_model()}: Remove listed models from the database

\item \code{add_models()}: add new models to database without removing existing models

}}
\examples{
\dontrun{
set_models(stardog, list(model1=c("http://example.com/graph1a", "http://example.com/graph1b"),
  model2="http://example.com/graph2"))
  }


}
