% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starmap.R
\name{snakeCamel}
\alias{snakeCamel}
\title{Take a field name and convert to snake case}
\usage{
snakeCamel(var)
}
\arguments{
\item{var}{the string to convert to came case}
}
\value{
the string in camel case
}
\description{
Take a field name and convert to snake case
}
