% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starmap.R
\name{snakePascal}
\alias{snakePascal}
\title{Take a field name and convert to snake case}
\usage{
snakePascal(var)
}
\arguments{
\item{var}{the string to convert to pascal case}
}
\value{
the string in pascal case
}
\description{
Take a field name and convert to snake case
}
