% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{Stardog}
\alias{Stardog}
\title{Create a stardog class (S3)}
\usage{
Stardog(
  endpoint = "http://localhost:5820",
  username = "admin",
  password = "admin",
  encoding = "UTF-8"
)
}
\arguments{
\item{endpoint}{The Stardog endpoint: a URL with the connection port (5820)}

\item{username}{Username for connecting to Stardog}

\item{password}{The Password}

\item{encoding}{Encoding to be used to parse responses. Defaults to UTF-8.}
}
\value{
The name of the S3 Stardog class. All functions in this package have this object as an argument.
}
\description{
This is the connection object for using the Stardog API. It
requires a Stardog endpoint and basic authentiation (username and password).
The database can be supplied as an option. There is a function that allows
the creation of a database and it will fill in the database slot.
An empty transaction slot is created for holding transaction id's.
}
\examples{
Stardog("http://localhost:5820", "admin", "admin")

}
