% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stardogR.R
\docType{package}
\name{stardogR}
\alias{stardogR}
\title{stardogR: a package to wrap the Stardog HTTP API}
\description{
stardogR enables queries, creating and dropping databases, adding triples, importing data from
data frames, managing namespaces and named graphs.
}
