% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starmap.R
\name{starmap}
\alias{starmap}
\title{Create a SMS mapping and a simple ontology}
\usage{
starmap(formulae, data, prefix = "", urn = "http://stardog.com/")
}
\arguments{
\item{formulae}{A vector of strings. Each string is a simple starmap formula}

\item{data}{The data frame to be mapped}

\item{prefix}{A sparql prefix}

\item{urn}{The prefix to each iri node written in full}
}
\value{
A list containing the ontology and the SMS mapping as strings
}
\description{
The terms of the formula must be column names of the data.frame data
The mappings and ontology are output as a list of two strings.
}
\details{
The formulae are based on the R formula language.
\code{x ~ y} maps x and y as iri nodes. there is an object property from x to y
\code{x ~ pref ~ y} max x and y as iri nodes. Bilds the object property from pref and y.
\code{x | v1 + v2} maps x as an iri node, with datatype properties v1 and v2.
The types of v1 and v2 are deduced from their datatypes in the dataframe.
\code{x | v1:v2} maps all columns from v1 through v2
\code{x | 1:4} maps columns 1 through 4.
Formula fragments can be combined, as in \code{x | v1 + v2 ~ y | w1 + w2 }
}
