% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{update_options}
\alias{update_options}
\title{Update options}
\usage{
update_options()
}
\value{
Vector of parameters
}
\description{
Options for update queries
}
\details{
This is a wheeze for providing constants to the rest of stardogR. See the
documentation for query_options()
}
