% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catch.R
\name{update_query}
\alias{update_query}
\title{Run an update query}
\usage{
update_query(stardog, q, ...)
}
\arguments{
\item{stardog}{Stardog object}

\item{q}{The update query}

\item{...}{Additional parameters for the query}
}
\value{
Success message for the request
}
\description{
Run an update query
}
