% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stored_query_fns.R
\name{update_stored}
\alias{update_stored}
\title{Update a stored query}
\usage{
update_stored(
  stardog,
  storedQuery = "CountEm",
  q = "select (count(distinct(?p)) as ?n) {?s ?p ?o .}"
)
}
\arguments{
\item{stardog}{Stardog object}

\item{storedQuery}{The name of the query to be updated}

\item{q}{The new query to be stored.}
}
\description{
Update a stored query
}
\details{
This function basically replaces the stored query with new code.
}
