% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starmap.R
\name{upperCaseFirst}
\alias{upperCaseFirst}
\title{Make the first letter of the string upper case}
\usage{
upperCaseFirst(var)
}
\arguments{
\item{var}{a string}
}
\value{
the string with the first character in upper case, leaving the rest as is
}
\description{
Make the first letter of the string upper case
}
