% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{fix_punctuation}
\alias{fix_punctuation}
\title{deal with commas and carriage returns}
\usage{
fix_punctuation(x)
}
\arguments{
\item{x}{A character string}
}
\value{
Returns the string with escaped quotes
}
\description{
Puts escaped quotes around strings that contain commas or carriage returns.
intended to imitate the behaviour of format_delim and prevent problems
when these symbols are given syntactic meaning.
}
