% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{query}
\alias{query}
\title{Query Stardog}
\usage{
query(
  stardog,
  q = "select (count(*) as ?n) {?s ?p ?o .}",
  namedGraph = NA,
  ...
)
}
\arguments{
\item{stardog}{A stardog object for the connect, containing a database}

\item{q}{The query expressed as a string}

\item{namedGraph}{Run the query against this named graph}

\item{...}{Options for the query. See Details}
}
\description{
Issues a query to Stardog and return the results in a dataframe.
}
\details{
Additional parameters:

- limit (integer) number of results to return

- reasoning (boolean) turn on reasoning. Not needed if a schema is included.

- schema (string) name of the reasoning schema

- offset (integer) number of results to skip before returning rows.
}
