% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_fns.R
\name{query}
\alias{query}
\title{Wrapper for specific query functions ask, construct and select}
\usage{
query(
  stardog,
  q = "select (count(*) as ?n) {?s ?p ?o .}",
  graph = NA,
  pretty = TRUE,
  ...
)
}
\arguments{
\item{stardog}{Stardog object}

\item{q}{An ask, construct, update or select query}

\item{graph}{runs the query against the specified named graph}

\item{pretty}{True when we want the select query to adjust the output with prefixes.}

\item{...}{Options for the query. See functions ask, construct and select}
}
\value{
the results of the query
}
\description{
Detects the type of query and runs the appropriate function
}
\details{
The graph parameter must be written out in full. "http://example.com/myGraph", say,
instead of "ex:myGraph". Stardog does not parse the prefix in this context.
}
