% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virtual_graphs.R
\name{register_databricks}
\alias{register_databricks}
\title{Register a data source on databricks}
\usage{
register_databricks(stardog, source_name, jdbc.url = "")
}
\arguments{
\item{stardog}{a stardog object}

\item{source_name}{a name for the data source}

\item{jdbc.url}{connection url}
}
\value{
The success code of the operation.
}
\description{
Registers the Hive tables from /user/hive/warehouse on Databricks
Allows connecting virtual graphs to these tables
}
\details{
There are instructions on Azure Databricks for forming jdbc.url. It
contains the address of the warehouse as well as an authentication token
for the user. Consequently, we don't need to supply username or password.
}
