% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_fns.R
\name{select}
\alias{select}
\title{Select query against Stardog}
\usage{
select(
  stardog,
  q = "select (count(*) as ?n) {?s ?p ?o .}",
  graph = NA,
  pretty = TRUE,
  raw = FALSE,
  ...
)
}
\arguments{
\item{stardog}{A stardog object for the connect, containing a database}

\item{q}{The query expressed as a string}

\item{graph}{Run the query against this named graph.}

\item{pretty}{description TRUE to replace the base iri with prefixes, where possible.}

\item{raw}{Boolean. Determines the format of the returned data. Raw=FALSE returns a dataframe}

\item{...}{Options for the query. See Details}
}
\value{
a dataframe containing the results.
}
\description{
Issues a select query to Stardog and return the results in a dataframe.
}
\details{
The graph parameter needs to be written out as a full IRI, say
"http://example.com/myGraph" rather than ex:myGraph.
Additional parameters:

- limit (integer) number of results to return

- reasoning (boolean) turn on reasoning. Not needed if a schema is included.

- schema (string) name of the reasoning schema

- offset (integer) number of results to skip before returning rows.
}
