% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adding_data_fns.R, R/query_fns.R
\name{update_query}
\alias{update_query}
\title{Run an update query}
\usage{
update_query(stardog, q, graph = NA, ...)

update_query(stardog, q, graph = NA, ...)
}
\arguments{
\item{stardog}{Stardog object}

\item{q}{The update query}

\item{graph}{Default graph for the "where" portion of the query}

\item{...}{Additional parameters for the query}
}
\value{
Success message for the request

Success message for the request
}
\description{
Run an update query

Run an update query
}
